Rename transformer -> parser
