+++ caldir-provider-google/src/commands/list_events.rs
@@ -7,9 +7,8 @@ use serde::Deserialize;
 
 use crate::DEFAULT_CALENDAR_ID;
 use crate::config;
-use crate::google_auth::get_valid_tokens;
-use crate::google_auth::redirect_uri;
-use crate::transformer::from_google_event;
+use crate::google_auth::{get_valid_tokens, redirect_uri};
+use crate::parser::from_google_event;
 
 #[derive(Debug, Deserialize)]
 struct ListEventsParams {
